+++ pyespn/classes/standings.py
@@ -1,9 +1,11 @@
-from pyespn.utilities import fetch_espn_data, get_an_id
+from pyespn.core.decorators import validate_json
+from pyespn.utilities import fetch_espn_data
 from pyespn.classes.player import Player
 from pyespn.classes.team import Manufacturer
 from pyespn.classes.stat import Record
 
 
+@validate_json("standings_json")
 class Standings:
     """
     Represents the standings for a racing league, including athletes, manufacturers, and performance records.
@@ -37,7 +39,7 @@ class Standings:
         """
 
         self.standings_json = standings_json
-        self.espn_instance = espn_instance
+        self._espn_instance = espn_instance
         self.standings = []
         self._load_standings_data()
 
@@ -74,15 +76,15 @@ class Standings:
             if 'athlete' in competitor:
                 athlete_content = fetch_espn_data(competitor.get('athlete', {}).get('$ref'))
                 this_athlete = Player(player_json=athlete_content,
-                                      espn_instance=self.espn_instance)
+                                      espn_instance=self._espn_instance)
             elif 'manufacturer' in competitor:
                 manufacturer_content = fetch_espn_data(competitor.get('manufacturer', {}).get('$ref'))
                 this_manufacturer = Manufacturer(manufacturer_json=manufacturer_content,
-                                                 espn_instance=self.espn_instance)
+                                                 espn_instance=self._espn_instance)
             records = []
             for record in competitor.get('records', []):
                 records.append(Record(record_json=record,
-                                      espn_instance=self.espn_instance))
+                                      espn_instance=self._espn_instance))
             full_athlete = {
                 'athlete': this_athlete,
                 'manufacturer': this_manufacturer,
@@ -90,3 +92,19 @@ class Standings:
             }
 
             self.standings.append(full_athlete)
+
+    @property
+    def espn_instance(self):
+        """
+            PYESPN: the espn client instance associated with the class
+        """
+        return self._espn_instance
+
+    def to_dict(self) -> dict:
+        """
+        Converts the Standings instance to its original JSON dictionary.
+
+        Returns:
+            dict: The standings's raw JSON data.
+        """
+        return self.standings_json